Add offset in dex frame maps.
Chrome can map in a dex file from an apk, which then makes
it strange to see a dex frame coming an apk. This cl will
set the elf_start_offset for the associated MapInfo object
so a dex pc frame can have an offset set to indicate where
the dex file came from within the apk.

Add new unit test for this behavior.

Bug: 356896129

Test: All unit tests pass.
Change-Id: I3a75ea1bb2052e7783ed92934173442b96f7c582