Import Cronet 131.0.6738.0 into AOSP
This commit brings in Cronet from Chromium 131.0.6738.0 (plus cherry-picks) into
AOSP, along with a few manual changes on top to make it work.

Note this commit was produced using a brand new kind of import workflow, which
is very different from previous imports. Most notably:

 - The gn2bp process is now driven from the Chromium side instead of the AOSP
   side
 - The 3-way merge is driven by copybara instead of merging across AOSP
   upstream-import and upstream-staging git branches (which are now de facto
   deprecated)

WHAT THIS COMMIT CONTAINS

The code in this commit was imported from:

    https://chromium.googlesource.com/experimental/chromium/src/+/refs/wip/cronet/aosp-131.0.6738.0
    Commit SHA: adfbec5f02e29d58e8ee246f0eaa6b62efa2b3b9

This commit is a descendant of Chromium src tag 131.0.6738.0 (369c6304), with
the following commits added on top:

 - 72 commits which are backported from the Chromium src main branch;
 - 2 commits that are not yet on the Chromium src main branch, but will be soon;
 - 4 commits that are specific to 131.0.6738.0 and are not going to be submitted
   to the Chromium src main branch.

On top of the imported code, various manual adjustments have been made to make
the code work:

 - Resolution of 3-way merge conflicts caused by divergence of the AOSP code
   relative to the Chromium code
 - Making various methods public to resolve package-private visibility issues,
   which only arise due to the use of jarjar in AOSP
 - Add cronet_aml_net_android_dummy_spnego_authenticator_java__testing to
   cronet_for_testing_net_java, otherwise
   net::HttpAuthHandlerNegotiateTest::SetUp crashes (see also
   https://crrev.com/c/6182156)

HOW TO REPRODUCE THIS COMMIT

The contents of this commit (minus the manual adjustments) can be reproduced by
running:

    gclient config --spec 'solutions = [
    {
        "name": "src",
        "url": "https://chromium.googlesource.com/experimental/chromium/src.git",
        "managed": False,
        "custom_deps": {},
        "custom_vars": {
            "checkout_copybara": True,
        },
    },
    ]
    target_os = ["android"]
    ' &&
    gclient sync --rev=refs/wip/cronet/aosp-131.0.6738.0 --shallow &&
    cd src &&
    vpython3 components/cronet/gn2bp/run_gn2bp.py

When running this command on top of this commit, the result is expected to be an
empty diff.

Test: TreeHugger
Fixes: 381421702
Change-Id: I8cb75517759f8d16923c9be48835cd0f2f6722b3