lib: rust_support: add interrupt save/restore bindings
This adds `interrupt_{save,restore}` bindings
with an `InterruptState` and `InterruptFlags` wrapper.

`ConstInterruptFlags` is the const generic ZST version of `InterruptFlags`,
with convenient type aliases in `interrupts::flags`.

`InterruptFlags::check` is added to ensure deadlocks don't happen
with the wrong combination of trying to disable IRQs and FIQs
vs. what's currently disabled.

Bug: None
Test: build.py
Change-Id: I5d4e08a3a03cfcbe6d5badc3193e31748ddbf31d