Autoupdate of 2 submodules
*build/blueprint updated to f6d588c5f0009d9f3a7795b4b17f6337c4424857
Convert TransitionMutators to use TransitionInfo

Make the configuration value passed by TransitionMutator be a
struct that implements a TransitionInfo interface.  The TransitionInfo
object replaces the variation strings returned by Split, passed in
and out of OutgoingTransition and IncomingTransition, and passed to
Mutate.

A TransitionInfo object still needs to be convertable to a variation
string (via it's Variation method) in order to be used in the path
to the module's intermediates directory and in human-readable error
messages.

When adding a dependency with an explicit variation after the transition
mutator has run, the variation string has to be converted to a
TransitionInfo object via TransitionMutator.TransitionInfoFromVariation.
If the TransitionInfo is too complex to store in a single string then
the TransitionInfoFromVariation can panic and adding dependencies with
explicit variations after the transition mutator has run is unsupported.

Bug: 372543712
Test: transition_test.go
Change-Id: I7545c72e95514172f27b8a4264666bfe10c642c6

*build/soong updated to d1516588b2cc78d6f333fc9e80e9cafe1d5bbaf1
Merge "Adapt Soong mutators to blueprint.TransitionInfo" into main