Refactor `ImTable::GetImtIndex()`.
Make it easier to calculate the IMT index without an actual
`ArtMethod` in preparation for a significant refactoring of
`ClassLinker::LoadClass()`.

Update documentation of `ArtMethod::imt_index_` which is
actually unused for abstract non-interface methods.

Also reorder code in `ClassLinker` without any modification
to make it easier to compare the differences in the upcoming
refactoring.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 329196666
Change-Id: Ic71ff323ea92325958d59536e83ab70d057b94df