ui: Scroll 'center' mode when scrolling tracks into view
Currently we use 'nearest' mode when scrolling tracks into view with
`scrollIntoView`. This doesn't account for sticky headers however, so
sometimes the track in question is actually still hidden behind one of
these sticky headers.

This CL uses 'center' mode instead which brings the track into the
very center of the viewport, which is much less likely to be behind
a sticky header.

Change-Id: I424055080cde0815abb7cfcfd70332617f93876d