test-map: Avoid log spam after running log-driver test
logd and the trusty log driver has bugs that does not cause a test
failure but eventually triggers an infinite loop that prints
error messages to the kernel log.

Stop logd while running the log-driver test to avoid this problem until
these bugs are fixed and logd stops reading from the device when the
driver tries to remove the device.

Bug: 397453308
Change-Id: If3834805afcac3665b228cd165834f4f9fbedce0