androidmk_denylist: ignore prebuilt Android SDK
The prebuilt SDK is used for non-platform Gradle builds, such as AndroidX and Android Automotive OS's Unbundled apps. The directory hasn't been a problem in the past, however the AAOS Unbundled build is adding the Android NDK within that directory (ag/31968665). The directory was chosen to match the location of the NDK in `androidx-main`.

Bug: b/395623570

Change-Id: I979f0c5705ff7e575100f0b2c9b002a84895883c